+++ app/src/main/java/woowacourse/movie/domain/seat/Seat.kt
@@ -1,6 +1,4 @@
-package woowacourse.movie.domain
-
-import SeatPosition
+package woowacourse.movie.domain.seat
 
 class Seat(val position: SeatPosition, val rank: SeatRank) {
