+++ app/src/main/java/woowacourse/movie/domain/seat/SeatRow.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.seat
 
 enum class SeatRow {
     A, B, C, D, E;
